FIX: dont set EEG annotations to ET object
This was the source of the bug where we were epoching to data that werent synced
to the flash times. While both of the EEG and ET raws are the same length after mne realign_raw and thus you should be able to epoch the ET objet to the EEG events, the original EEG raw and ET raw had different orig_times and so setting EEG annots to ET raw will put them in the wrong place. This is a documented limiation of annotations. if t0 is not 0, annotations can behave weird.